？

+++ src/po/GoodsPO.java
@@ -32,26 +32,26 @@ public class GoodsPO implements Serializable {
     /**
      *进价
      */
-    private int purchasePrice;
+    private double purchasePrice;
     /**
      * 零售价
      */
-    private int retailPrice;
+    private double retailPrice;
     /**
      *最近进价
      */
-    private int recentPurPrice;
+    private double recentPurPrice;
     /**
      **最近零售价
      */
-    private int recentRetPrice;
+    private double recentRetPrice;
 
     public GoodsPO() {
         super();
     }
 
-    public GoodsPO(String name, String type, long commodityNum, int purchasePrice, int retailPrice,
-            int recentPurPrice, int recentRetPrice) {
+    public GoodsPO(String name, String type, long commodityNum, double purchasePrice, double retailPrice,
+                   double recentPurPrice, double recentRetPrice) {
         this.name = name;
         this.type = type;
         this.commodityNum = commodityNum;
@@ -61,8 +61,8 @@ public class GoodsPO implements Serializable {
         this.recentRetPrice = recentRetPrice;
     }
 
-    public GoodsPO(int number, String name, String type, long commodityNum, int purchasePrice, int retailPrice,
-                   int recentPurPrice, int recentRetPrice) {
+    public GoodsPO(int number, String name, String type, int commodityNum, double purchasePrice, double retailPrice,
+                   double recentPurPrice, double recentRetPrice) {
         super();
         this.number = number;
         this.name = name;
@@ -111,35 +111,35 @@ public class GoodsPO implements Serializable {
         this.commodityNum = commodityNum;
     }
 
-    public int getPurchasePrice() {
+    public double getPurchasePrice() {
         return purchasePrice;
     }
 
-    public void setPurchasePrice(int purchasePrice) {
+    public void setPurchasePrice(double purchasePrice) {
         this.purchasePrice = purchasePrice;
     }
 
-    public int getRetailPrice() {
+    public double getRetailPrice() {
         return retailPrice;
     }
 
-    public void setRetailPrice(int retailPrice) {
+    public void setRetailPrice(double retailPrice) {
         this.retailPrice = retailPrice;
     }
 
-    public int getRecentPurPrice() {
+    public double getRecentPurPrice() {
         return recentPurPrice;
     }
 
-    public void setRecentPurPrice(int recentPurPrice) {
+    public void setRecentPurPrice(double recentPurPrice) {
         this.recentPurPrice = recentPurPrice;
     }
 
-    public int getRecentRetPrice() {
+    public double getRecentRetPrice() {
         return recentRetPrice;
     }
 
-    public void setRecentRetPrice(int recentRetPrice) {
+    public void setRecentRetPrice(double recentRetPrice) {
         this.recentRetPrice = recentRetPrice;
     }
